requirements.txt
pydantic>=2.0.0
pyyaml>=6.0
jinja2>=3.1.0
click>=8.1.0
jsonschema>=4.0.0
black>=23.0.0
inflection>=0.5.1
typing-extensions>=4.5.0
